refactor(tanstack-start): Change structure of server subpath
