+++ packages/tanstack-start/src/server/index.ts
@@ -1,38 +1,4 @@
-import type { AnyRouter } from '@tanstack/react-router';
-import type { EventHandler } from 'vinxi/http';
-
-import { rootAuthBeforeLoader } from './rootAuthBeforeLoader';
-
-export type RequestHandler<TRouter extends AnyRouter> = (ctx: {
-  request: Request;
-  router: TRouter;
-  responseHeaders: Headers;
-}) => Promise<Response>;
-export type CustomizeRequestHandler<TRouter extends AnyRouter> = (cb: RequestHandler<TRouter>) => EventHandler;
-
-export function clerkMiddlewareHandler<TRouter extends AnyRouter>(eventHandler: CustomizeRequestHandler<TRouter>) {
-  return (cb: RequestHandler<TRouter>): EventHandler => {
-    return eventHandler(async ({ request, router, responseHeaders }) => {
-      try {
-        const authContext = await rootAuthBeforeLoader(request);
-
-        router.update({
-          context: authContext || {},
-        });
-
-        await router.load();
-      } catch (error) {
-        if (error instanceof Response) {
-          return error;
-        }
-
-        throw error;
-      }
-
-      return cb({ request, router, responseHeaders });
-    });
-  };
-}
+export * from './middlewareHandler';
 
 /**
  * Re-export resource types from @clerk/backend
