+++ packages/tanstack-start/tsconfig.json
@@ -11,7 +11,6 @@
     "esModuleInterop": true,
     "jsx": "react-jsx",
     "forceConsistentCasingInFileNames": true,
-    "verbatimModuleSyntax": true,
     "isolatedModules": true,
     "lib": ["DOM", "DOM.Iterable", "ES2022"],
     "module": "ES2022",
